requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-filter==23.3
django-recaptcha==3.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
weasyprint==60.1
whitenoise==6.6.0
Pillow==10.1.0  # Required for image handling
